requirements.txt
Flask>=2.1
Flask-WTF>=1.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
python-dotenv>=1.0
gunicorn>=20.0
psycopg2-binary>=2.9
